app/pyproject.toml
[tool.poetry]
name = "zebra3"
version = "0.1.0"
description = ""
authors = ["giannifer7 <gianni.ferrarotti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Kivy = "^2.1.0"
kivymd = "^0.104.2"
pyjnius = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
